play 2.1 and scala 2.10 migration (not tested, warnings remain)

+++ Play20/framework/src/play/src/main/scala/play/core/server/Server.scala
@@ -11,9 +11,8 @@ import play.api.libs.concurrent._
 import akka.actor._
 import akka.actor.Actor._
 import akka.routing._
-import akka.config._
 import akka.pattern.Patterns.ask
-import akka.util.duration._
+import scala.concurrent.duration._
 import akka.util.Timeout
 
 trait WebSocketable {
@@ -26,11 +25,11 @@ trait WebSocketable {
  */
 trait Server {
 
-  // First delete the default log file for a fresh start
+  // First delete the default log file for a fresh start (only in Dev Mode)
   try {
-    scalax.file.Path(new java.io.File(applicationProvider.path, "logs/application.log")).delete()
+    if (mode == Mode.Dev) scalax.file.Path(new java.io.File(applicationProvider.path, "logs/application.log")).delete()
   } catch {
-    case _ =>
+    case _: Exception =>
   }
 
   // Configure the logger for the first time
@@ -38,14 +37,9 @@ trait Server {
     Map("application.home" -> applicationProvider.path.getAbsolutePath),
     mode = mode)
 
-  // Start the main Invoker
-  val invoker = Invoker(applicationProvider)
-
-  // store the invoker in a global variable
-  Invoker.init(invoker)
-
   val bodyParserTimeout = {
-    Configuration(Invoker.system.settings.config).getMilliseconds("akka.actor.retrieveBodyParserTimeout").map(_ milliseconds).getOrElse(1 second)
+    //put in proper config
+    1 second
   }
 
   def mode: Mode.Mode
@@ -61,7 +55,7 @@ trait Server {
           (maybeAction.getOrElse(Action(BodyParsers.parse.empty)(_ => application.global.onHandlerNotFound(request))), application)
         }
       } catch {
-        case e => Left(e)
+        case e: Exception => Left(e)
       }
     }
 
@@ -70,11 +64,11 @@ trait Server {
       Logger.error(
         """
         |
-        |! %sInternal server error, for request [%s] ->
+        |! %sInternal server error, for (%s) [%s] ->
         |""".stripMargin.format(e match {
           case p: PlayException => "@" + p.id + " - "
           case _ => ""
-        }, request),
+        }, request.method, request.uri),
         e)
 
       DefaultGlobal.onError(request, e)
@@ -91,20 +85,9 @@ trait Server {
 
   }
 
-  def invoke[A](request: Request[A], response: Response, action: Action[A], app: Application) = {
-    invoker.actionInvoker ! Invoker.HandleAction(request, response, action, app)
-  }
-
-  def getBodyParser[A](requestHeaders: RequestHeader, bodyFunction: BodyParser[A]): Promise[Iteratee[Array[Byte], Either[Result, A]]] = {
-    val future = ask(invoker.actionInvoker, Invoker.GetBodyParser(requestHeaders, bodyFunction), bodyParserTimeout)
-    future.asPromise.map(_.asInstanceOf[Iteratee[Array[Byte], Either[Result, A]]])
-  }
-
   def applicationProvider: ApplicationProvider
 
   def stop() {
-    Invoker.uninit()
-    invoker.stop()
     Logger.shutdown()
   }
